Fixes bug that resulted in status bubble getting hidden
Interesting enough this was happening because of a focus bug. Turns
out a FocusController associated with one DesktopNativeWidgetAura
could attempt to focus an aura::Window in a different
DesktopNativeWidgetAura. When this happened it would move the
content_window on top of the status bar.

We should only activate windows within the same root.

BUG=345698
TEST=covered by unit test, see bug as well.
R=ben@chromium.org

Review URL: https://codereview.chromium.org/186733011