Automate rolling of Blink DEPS through the Chromium CQ
This bot is very simple it:
1. Checks to see if there is an "active" roll CL.
2. If that roll is failed/timed out (8hrs) closes it, otherwise sleeps.
3. If there is no more active roll it opens a new one, CCing the gardeners.

See an example roll:
https://codereview.chromium.org/15255007

This is just moving the code from:
https://codereview.chromium.org/15539006
out of Blink and into the tools repository.

This script has been running successfully since May
and has become a critical part of the Blink infrastructure
and thus has no business running on my personal machine at home. :)

This has hard-coded paths for my machine, but someone
who wishes to set this up as a real cron job should change
those to whatever is correct for the cron machine.

Hopefully the style is sufficiently close that someone
who actually owns this repository can amend it to their liking.

BUG=242461

Review URL: https://codereview.chromium.org/66673006