Define print format macros for NSInteger & NSUInteger
The size of NSInteger and NSUInteger varies between 32-bit and 64-bit
architectures, however does not provides macro to safely format them
and instead recommend casting the value to the larger version that is
used on 64-bit architecture.

Using a cast could cause some formatting to be missed (if the type of
a variable changes), so instead we define our own macros to format those
types safely.

BUG=349458

Review URL: https://codereview.chromium.org/187793003