Make a graph tool of dmprof useful.
1. Change the graph from bar-chart to area chat.
Now x-axis shows time line.  It is instinctive.

2. Enable to choose policies in GUI.
It is a waste of time to apply many policies and make different HTML files.
Now it can be done with clicks.  No need to specify a policy in converting json to HTML.

Usage:
$ dmprof /path/to/first.heap > temp.json
$ graph.py temp.json > graph.html  # no need to specify a poilcy
$ chrome graph.html    # open the HTML file in Chrome


BUG=None
NOTRY=true

Review URL: https://codereview.chromium.org/189623009