Support the Aec dump for the APM in chrome.
The code will run behind the kEnableAudioTrackProcessor flag and the behaviour will be compatible with the existing behaviours.

Since AEC dump can be started only once, so we can't re-create a new instance of MediaStreamAudioProcessor when the format changes. This CL will create the MediaStreamAudioProcessor in the constructor of WebRtcAudioCapturer and call OnCaptureFormatChanged() to the MediaStreamAudioProcessor when the format changes.

BUG=264611
TEST=manual enable aec dump in chrome://webrtc-internals and verify the aec dump audio files.

Review URL: https://codereview.chromium.org/187913002