Adding a gray semi transparent backdrop behind the topmost window within the default container
This is part of the "always maximized" feature. Windows which cannot be maximized - or cannot be made to cover the entire screen should have a backdrop behind them which covers the desktop.

Tried various ways to implement this and this seems to be the best solution.

BUG=337567, 337563

Review URL: https://codereview.chromium.org/169643005