Remove --purge-memory-button.
This was added years ago as a way to test the memory purger, but we never ended
up using the memory purger.  This removes the button and the MemoryPurger class,
but not all the scattered pieces of memory-purging functionality the class could
call.  Those will be removed in subsequent CLs.

BUG=350455
TEST=none
R=sky@chromium.org

Review URL: https://codereview.chromium.org/192573010