[Telemetry] OS X: Properly clean up power measurement
In the page cycler measurement, the call that verifies that a page has finished loading could throw a timeout exception causing the power metric to not be properly shut down.

This can cause many instances of powermetric to run in parallel.

This patch adds additional cleanup code to the powermetric to make sure the backing utility is well dead.

BUG=339162

Review URL: https://codereview.chromium.org/192243002