cc: Don't try to reduce main thread latency in smoothness takes priority mode
If the compositor is in smoothness takes priority mode, do not try to
skip BeginMainFrames to reduce main thread latency because the two goals
are mutually exclusive. This fixes the problem of constantly going in
and out of low latency mode on pages where the DOM is constantly
modified in response to (impl-side) scrolling.
    
TEST=The green box on http://jsbin.com/sifisogo/11?quiet=1 should not
     move choppily (but may still move out of sync with the red box.)
BUG=309648

Review URL: https://codereview.chromium.org/188753002