Fix LayerTreeImpl::VerticalAdjust().
The correct test of whether a scrollbar layer needs to receive vertical
adjust or not is whether its clip layer is the
InnerViewportContainerLayer. This CL perfroms that test, and modifies
the signature of LayerTreeImpl::VerticalAdjust() to take a layer id
(int) instead of a LayerImpl*.

This should only have any effect when  both inner and outer pinch viewports
are active.

Review URL: https://codereview.chromium.org/192763002