Closing IndexedDBConnection if still open in IndexedDBContextImpl::ConnectionClosed()
If the WebIDBDatabaseImpl destructor closes an open connection then
this will result in the IndexedDBFactory having a dangling pointer
to a deleted IndexedDBDatabase instance - because that is only removed
when database is explicitly closed via IndexedDBDatabase::Close().

To protect against this the connection is closed, which in turn
informs the IndexedDBFactory so that it can remove references.

BUG=350100

Review URL: https://codereview.chromium.org/189843011