content_shell: Add event filtering to PluginObject
For testing input event filtering in plugins, this change adds a
"alwaysFilterEvents" argument to PluginObject. That argument causes the plugin
to always return a non-zero value from HandleEvent, which requests the event be
"filtered" to stop propagating it to other handlers on the page.

BUG=324351

Review URL: https://codereview.chromium.org/191383005