Make ant failures easier to debug
Normally, ant commands (even when successful) have lots of useless
output. So, for most of our ant commands we pass the -quiet flag. This
is much better for successful builds but it makes debugging failures
more difficult (though even without -quiet ant failures are difficult to
diagnose without -verbose.

So, the only way that ant failures are easy to debug is if the -verbose
flag is passed to ant. This flag is much too noisy to always pass. This
change makes it so that if a call to ant fails, we repeat that call with
the -verbose flag (and without -quiet) so that the output that the
developer sees is most useful.

Review URL: https://codereview.chromium.org/193283002