Reenable the ResourceBundleImageTest.GetImageNamed test for Windows.
The test was failing due to the change to load 200 px resources for Windows for high dpi
support. Reason being the test expected that the scale factor for images on platforms other
than Chrome OS was 100 px which is not the case anymore.

Updated the test to have similar expectations on Windows.

BUG=349607
R=sky@chromium.org
TBR=sky

Review URL: https://codereview.chromium.org/193713003