Do not set RootView's gesture handler on last ET_GESTURE_END
There might be cases that a RootView is trying to dispatch an
ET_GESTURE_END event and no gesture handler is set for that. In such
cases, after the event is handled by a view, it should not set that view
as the gesture handler for the RootView, since that gesture sequence is
finished.

BUG=334968

Review URL: https://codereview.chromium.org/190663017