Add BluetoothDevice::VendorIDSource()
The Vendor ID of a device can be either allocated by the Bluetooth SIG
or USB IF, providing two completely seperate overlapping namespaces for
identifiers.

In order to distinguish which vendor an identifier represents, add a
VendorIDSource() method to BluetoothDevice that returns BLUETOOTH or
USB respectively.

Add to the metrics collection as well.

BUG=350432
TEST=device_unittests
R=armansito@chromium.org, isherman@chromium.org

Review URL: https://codereview.chromium.org/191223003