Require SSE2 for all 32-bit Linux builds and remove disable_sse2 option.
Previously, only 32-bit Google Chrome Linux was being built with x87 floating point math, which led to bugs caused by 80-to-{32,64}-bit rounding errors that weren't caught on the main Chromium builders. UMA data tells us there are ~0 Google Chrome Linux users without SSE2, making a strong case to require SSE2 across the board and align the Google Chrome build closer to the Chromium build.

As for disable_sse2 option, it was added in r45777 but it's unclear if anyone is using it outside of some other GYP code that flips it to 1 for ARM and MIPS architectures. Instead replace the few instances that don't support SSE2 with specific checks for the target architecture and remove the option entirely.

BUG=348761
R=jamesr@chromium.org
TBR=cpu

Review URL: https://codereview.chromium.org/187423002