Accept remote urls as gospel in git_cache.
git_cache rewrites all urls to https://netloc/path when
setting them as remote urls. This doesn't work when the
remote is actually local on disk, and the "url" is a
/path/to/repo. This seems like an unlikely case, but it
is leveraged by testing, and it doesn't hurt to support
it, so now we do.

R=iannucci@chromium.org
BUG=261619

Review URL: https://codereview.chromium.org/196493002