Add ViewHostMsg_DocumentDetached to SwappedOutMessages::CanSendWhileSwappedOut.
ViewHostMsg_DocumentDetached must be sent from the renderer process when the user leaves the page which is running SharedWorker.
But it is not sent in the current code when the user is going to another domain page.
It is because ViewHostMsg_DocumentDetached is not handled in SwappedOutMessages::CanSendWhileSwappedOut.

BUG=327256

Review URL: https://codereview.chromium.org/194443002