Remove Profile knowledge from MutableProfileOAuth2TokenService
This CL removes all remaining calls to profile() in
MutableProfileOAuth2TokenService, changing them to be calls on the
SigninClient. Concretely, it adds the following APIs to SigninClient:
- CanRevokeCredentials()
- GetURLRequestContext()

Note that the latter is a BrowserContext-level API and thus could go in a
driver within the Signin component. However, at the current time it does not
appear that there will be any other reason that the Signin component should be
a layered component; I would like to avoid adding the extra complexity in
internal structure until there is a compelling need for introducing the
layering.

BUG=334184

Review URL: https://codereview.chromium.org/192373010