CoreAnimation: Ensure that frames are drawn after window resize
In this bug, the layer will be displayed in a strange state where the
the GL contents are inappropriately translated, and the background
color, which should never be visible for a CAOpenGLLayer sub-class, is
visible.

There are two bugs at play.

The first bug is in drawInCGLContext where we infer the size of the
surface we're drawing to from the layer's bounds. This is not correct,
the size of the surface being drawn to is the GL_VIEWPORT when
drawInCGLContext is called. Re-cast this as a window size, taking
into account contents scale (this can be simplified when the
non-CoreAnimation code is not needed).

With that fixed, we no longer see strange garbage. Rather, we start
seeing just the old frame (so, the bottom bit of the frame corresponding
to size of the now-removed infobar is now missing).

The second bugs is another instance of the well-established fact that
setDisplay will often be dropped on the floor. This happens pretty
consistently during the resize. Add the usual fix of immediately
calling displayIfNeeded.

BUG=350817

Review URL: https://codereview.chromium.org/194593003