Make RenderWidgetHostViewCocoa return YES to mouseDownCanMoveWindow. (Mac)
A point in a window is draggable if its responder chain all return YES
to mouseDownCanMoveWindow. To implement "-webkit-app-region:
drag | no-drag", we have a WebContentsViewMac that returns NO by
default. For frameless app windows, we make it return YES, and add
subviews that return NO to exclude non-draggable regions.

Since RenderWidgetHostViewCocoa is also a subview of
WebContentsViewMac and covers the entire window, it needs to return
YES for dragging to work.

The default implementation of mouseDownCanMoveWindow returns
![self isOpaque], so it's likely than turning on core animation was the
cause of this bug.

BUG=350355

Review URL: https://codereview.chromium.org/194283002