Cast: Log compression and getRawEvents JS API change.
An additional bool argument in LogSerializer indicates the serialized
log should be compressed. Compression will be done with zlib, and output
will be in gzip format.

An additional bool argument in LogDeserializer indicates the input is
compressed and that a decompression should be performed before the
actual deserialization.

Changed LogSerializer API to only take data from 1 stream, since we are
intending to output separate files for each stream anyway.

CastSessionDelegate will always ask for logs compressed. Though it
can be changed in the future.

The sender app has been modified to allow the user to write logs to
file in compressed format.

The cast extension now works with base::BinaryValue rather than
a std::string. The V8ValueConverter converts it to an ArrayBuffer
object to JS. (A Blob can be constructed from it if desired) The JS
API has changed so that getRawEvents returns an ArrayBuffer
rather than a DOMString in the callback.

BUG=301920, 338574

Review URL: https://codereview.chromium.org/189583004