Fixed alignment issue in tray in guest mode.
In guest mode, the avatar in the system tray is replaced by a label. The label has a variable size vs the fixed height of avatar images. This causes the tray container to shrink resulting in alignment issues.

I've modified the border around the label so that the height of a |TrayUser| will be consistent. I've also add a unit test to enforce this.

BUG=340673
TEST=Current unit tests pass. New unit test passes. Verify bug 340673.

Review URL: https://codereview.chromium.org/189193003