Reland: Fix mojo_js_unittests
The original CL is: https://codereview.chromium.org/197463005/

Comparing with the original CL, this CL doesn't enable JSTest.codec on the bots. I realized that I have other issues to fix before I can re-enable it (the test depends on files in gen/ which are not copied to testing bots).

TBR=sky@chromium.org
BUG=351214
TEST=None

Review URL: https://codereview.chromium.org/195793019