cc: Run rasterize benchmark on the worker thread.
Since the change to reuse the original picture for one of the threads,
the rasterize_and_record_micro benchmark code is no longer correct, as
it is using the original picture to rasterize on the impl thread.

This patch changes the behaviour to run the benchmark on one of the
worker threads.

BUG=350684
R=reveman@chromium.org

Review URL: https://codereview.chromium.org/198053003