Attempting to reland the Overview Button Change.
The inital change was landed as 256420.
The change was reverted by 256535, as a Win 7 release test failed on the build server. Another change in ash was also reverted.
I was unable to reproduce the test failure on a local Win 7 machine.

"Revert of Add a new status area control for Overview Mode. This control is only available (https://codereview.chromium.org/180743016/)"

This reverts commit 65e11004319c51eb14787d51102a3db9100da480.

BUG=337590
TBR=oshima
R=flackr

Review URL: https://codereview.chromium.org/196473011