Ask whether to end session when last window is closed in public session
A user closing all browser and app windows may either be a power user
cleaning up or a user less familiar with Chrome OS trying to exit the
browser (and thus, the session). Since we do not know which one is the
case, show a dialog that asks the user whether to exit the session when
the last window is closed.

BUG=278247
TEST=New browser test
R=derat@chromium.org, stevenjb@chromium.org

Review URL: https://codereview.chromium.org/197333002