Presubmit: fix accepted copyright dates.
2006-2009/10 was a valid copyright date format before 2011 (when we decided to
use only one year, not a range).

The current directive with copyright is in fact to not change old copyright
dates, so there is no plan to go back and "fix" old files. As such, the
presubmit checks should not be complaining about old files.

BUG=none

Review URL: https://codereview.chromium.org/200643002