[android_webview] Increase the scroll distance in one of the tests.
The testNoSpuriousOverScrolls test was failing on the N5 because
of what seems to be a difference in the touch slop region size.
Increasing the number of pixels we scroll by in the test fixes
the issue but also caused a fling animation which is the reason
for the custom AwScrollOffsetManager.

BUG=351476

Review URL: https://codereview.chromium.org/198013003