Wire up forced re-enrollment in OOBE.
This changes the OOBE flow to wait for the auto-enrollment check to
complete. If successful, proceed to sign-in or enrollment depending on
the result of the check. If there's a network error, show the new
variant of the error screen telling the user that they need to connect
to a network on which we can make a successful connection to Google.

BUG=351663

Review URL: https://codereview.chromium.org/200313002