Set the name of several iframes in browser resources
Set the name of several iframes in browser resources in addition to their id,
to avoid breakage when Blink gets updated so that an iframe's name is no longer
set to its id when missing (as per specification).

Several scripts / browser tests rely on the iframe's name being the same.

R=arv, jochen
BUG=347169

Review URL: https://codereview.chromium.org/197763016