Ensure the DomDistillerViewerSource is registered at startup.
This changes the registration time of the DomDistillerViewerSource which
handles requests to the chrome-distiller:// scheme. Before this CL, the
URLDataSource was registered as soon as the BrowserContextKeyedService
was requested, which is too late when the user directly tries to distill a URL.
This changes it so that the scheme registration happens as soon as the
profile has been created.

Since the instantiation of the service causes the database for the
DomDistillerStore to load, an instance holder is created which causes the
service to be lazily created when it is needed.

BUG=319881

Review URL: https://codereview.chromium.org/188373002