Mac: Don't animate throbbers during window minimize animations.
Only animate the sprites if the view is attached to a window, and that window
is not currently minimized or in the middle of a minimize or deminimize
animation.
I don't know why it was causing the bad unminimize animation, but anyway
it's good not to animate when not visible.

BUG=350329

Review URL: https://codereview.chromium.org/200853003