Update frame color API to use unions.
This allows us to use the 'frame' option either as a string (for type)
or as a dictionary (for type and / or color).

Note, this is backwards incompatible with the previous way of setting frame color. To use frame color you now use 'frame' instead of 'frameOptions'.

BUG=339558

Review URL: https://codereview.chromium.org/197813008