Fix a deadlock by dropping frames in RtcVideoEncoder.
If there are no free input and output buffers, drop the frame to avoid a
deadlock. Dropping a frame is fine. The encoder has been filled with all input
buffers. If this is a key frame and Encode() returns error, WebRTC will request
a key frame again. Besides, webrtc will drop a frame if Encode() blocks too
long.

The caller of Encode() holds a webrtc lock. The deadlock happens when:
(1) Encode() is waiting for the frame to be encoded in EncodeOneFrame().
(2) There are no free input buffers and they cannot be freed because the
    encoder has no output buffers.
(3) Output buffers cannot be freed because ReturnEncodedImage is queued on
    libjingle worker thread to be run. But the worker thread is waiting for
    the same webrtc lock held by the caller of Encode().

BUG=349406
TEST=Run apprtc.appspot.com loopback on CrOS.
     Run apprtc.appspot.com between CrOS and Linux.

Review URL: https://codereview.chromium.org/199973003