Allows the guest session on CrOS to load extension resource.
The current implementation inhibits extensions to load resources if it's in incognito mode and the extension is not enabled for incognito mode.  That's right in a regular (non-guest) session.

However, the guest session is in incognito mode by default, and it doesn't have a regular (non-incognito) context.  So we should allow resource loading if it's the guest session, otherwise there is no way for component extensions to load resources.  Note component extensions are never listed in the extension list, so users cannot enable those extensions for incognito mode.

BUG=339318
TEST=Manually done.
TBR=davemoore@chromium.org

Review URL: https://codereview.chromium.org/182493004