Improve network error notifications and cleanup code
This CL does the following:
* Stores last_error in NetworkState
* Uses Service.PreviousError when Serice.Error is empty
* Relies on Service + NetworkState properties instead of awkwardly passing shill errors around (inconsistently)

Note: This CL was developed in parallel to PreviousError in Shill. Rather than revert the last_error changes, I decided to track both and LOG if we use one or the other under the assumption that more information is better.

BUG=333955,344398,337608,332620
R=armansito@chromium.org, pneubeck@chromium.org

Review URL: https://codereview.chromium.org/165263003