Fail determine_new_failures() if there are too many failures.
Prior to this patch, if we ran the blink recipe and *every* test failed,
the logic for determining new failures wouldn't work right (we'd get the
same list of failures both with and without the change).

This patch adds a failsafe setting, which is that we will consider the
patch as bad if there are more than 75 failures, period.

This is the wrong fix, and we need to add tests to get recipe_modules
test coverage back up, but this will do as a stop-gap.

TBR=phajdan@chromium.org, iannucci@chromium.org
BUG=353061

Review URL: https://codereview.chromium.org/201613007