Fixes the MessageCenterNotificationsTest with multi-profiles flag.
BasicAddCancel: the number of notification may change because of
multi-profile notification. Adds 'CancelAll' to clear it.

QueueWhenCenterVisible: the log of TestAddObserver also contains
the events from the multi-profile notification. To deal with this,
this CL separates the event log by the notification id and check
the expectation. This will simplify the expecation.

BUG=351225
R=dewittj@chromium.org
TEST=browser_tests --multi-profiles

Review URL: https://codereview.chromium.org/195993002